Refactor forms
