+++ enhydris/forms/__init__.py
@@ -1,0 +1,4 @@
+from .auth import MyRegistrationForm
+from .timeseries import DownloadDataForm
+
+__all__ = ("MyRegistrationForm", "DownloadDataForm")
