+++ enhydris/forms/timeseries.py
@@ -1,9 +1,6 @@
 from django import forms
 from django.utils.translation import ugettext_lazy as _
 
-from captcha.fields import CaptchaField
-from registration.forms import RegistrationFormTermsOfService
-
 from enhydris import models
 
 DOWNLOAD_FORMATS = [
@@ -39,13 +36,3 @@ class DownloadDataForm(forms.Form):
             raise forms.ValidationError("Invalid TimeseriesGroup")
         self.fields["timeseries_id"].choices = self.get_choices()
         return tgid
-
-
-class MyRegistrationForm(RegistrationFormTermsOfService):
-    captcha = CaptchaField(label=_("Are you human?"))
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.fields["username"].help_text = _(
-            "150 characters or fewer. Letters, digits and @/./+/-/_ only."
-        )
